+++ Sources/MereRunCLI/Support/WorkflowRunner.swift
@@ -22,215 +22,6 @@ struct WorkflowRunOutcome: Codable, Equatable {
     }
 }
 
-struct WorkflowProcessResult: Equatable {
-    let status: Int32
-    let stdout: String
-    let stderr: String
-    let terminationReason: WorkflowProcessTerminationReason
-
-    init(
-        status: Int32,
-        stdout: String,
-        stderr: String = "",
-        terminationReason: WorkflowProcessTerminationReason = .exit
-    ) {
-        self.status = status
-        self.stdout = stdout
-        self.stderr = stderr
-        self.terminationReason = terminationReason
-    }
-
-    var failureSummary: String {
-        var summary = terminationReason == .uncaughtSignal
-            ? "terminated by signal \(status)"
-            : "exited with status \(status)"
-        if !stderr.isEmpty {
-            summary += ". stderr: \(stderr)"
-        } else if !stdout.isEmpty {
-            summary += ". stdout: \(stdout.suffix(16 * 1_024))"
-        }
-        return summary
-    }
-}
-
-enum WorkflowProcessTerminationReason: String, Equatable {
-    case exit
-    case uncaughtSignal = "uncaught_signal"
-
-    init(_ reason: Process.TerminationReason) {
-        switch reason {
-        case .exit:
-            self = .exit
-        case .uncaughtSignal:
-            self = .uncaughtSignal
-        @unknown default:
-            self = .exit
-        }
-    }
-}
-
-private struct WorkflowProcessTimeoutError: LocalizedError {
-    let seconds: Int
-
-    var errorDescription: String? {
-        "Process timed out after \(seconds) seconds."
-    }
-}
-
-private final class WorkflowProcessTimeoutState: @unchecked Sendable {
-    private let lock = NSLock()
-    private var timedOut = false
-
-    func markTimedOut() {
-        lock.lock()
-        timedOut = true
-        lock.unlock()
-    }
-
-    var didTimeOut: Bool {
-        lock.lock()
-        defer { lock.unlock() }
-        return timedOut
-    }
-}
-
-enum WorkflowChildProcessRegistry {
-    static let directoryName = "worker-child-pids"
-    static let legacyFilename = "worker-child.pid"
-    private static let lock = NSLock()
-
-    static func register(
-        _ processID: Int32,
-        in runDirectory: URL,
-        fileManager: FileManager = .default
-    ) throws {
-        lock.lock()
-        defer { lock.unlock() }
-        let directory = runDirectory.appendingPathComponent(directoryName, isDirectory: true)
-        let entry = directory.appendingPathComponent("\(processID).pid")
-        do {
-            try fileManager.createDirectory(at: directory, withIntermediateDirectories: true)
-            try Data(String(processID).utf8).write(to: entry, options: .atomic)
-            try Data(String(processID).utf8).write(
-                to: runDirectory.appendingPathComponent(legacyFilename),
-                options: .atomic
-            )
-        } catch {
-            try? fileManager.removeItem(at: entry)
-            throw error
-        }
-    }
-
-    static func unregister(
-        _ processID: Int32,
-        in runDirectory: URL,
-        fileManager: FileManager = .default
-    ) {
-        lock.lock()
-        defer { lock.unlock() }
-        let entry = runDirectory
-            .appendingPathComponent(directoryName, isDirectory: true)
-            .appendingPathComponent("\(processID).pid")
-        try? fileManager.removeItem(at: entry)
-        let legacy = runDirectory.appendingPathComponent(legacyFilename)
-        if readProcessID(at: legacy, fileManager: fileManager) == processID {
-            try? fileManager.removeItem(at: legacy)
-        }
-    }
-
-    static func processIDs(
-        in runDirectory: URL,
-        fileManager: FileManager = .default
-    ) -> [Int32] {
-        let directory = runDirectory.appendingPathComponent(directoryName, isDirectory: true)
-        let entries = (try? fileManager.contentsOfDirectory(
-            at: directory,
-            includingPropertiesForKeys: [.isRegularFileKey, .isSymbolicLinkKey]
-        )) ?? []
-        var processIDs = Set(entries.compactMap { entry -> Int32? in
-            guard entry.pathExtension == "pid",
-                  let values = try? entry.resourceValues(forKeys: [.isRegularFileKey, .isSymbolicLinkKey]),
-                  values.isRegularFile == true,
-                  values.isSymbolicLink != true else {
-                return nil
-            }
-            return readProcessID(at: entry, fileManager: fileManager)
-        })
-        if let legacy = readProcessID(
-            at: runDirectory.appendingPathComponent(legacyFilename),
-            fileManager: fileManager
-        ) {
-            processIDs.insert(legacy)
-        }
-        return processIDs.sorted()
-    }
-
-    @discardableResult
-    static func terminateAll(
-        in runDirectory: URL,
-        fileManager: FileManager = .default
-    ) -> [Int32] {
-        let processIDs = processIDs(in: runDirectory, fileManager: fileManager)
-        for processID in processIDs {
-            _ = kill(processID, SIGTERM)
-        }
-        return processIDs
-    }
-
-    static func clear(
-        in runDirectory: URL,
-        fileManager: FileManager = .default
-    ) throws {
-        lock.lock()
-        defer { lock.unlock() }
-        let legacy = runDirectory.appendingPathComponent(legacyFilename)
-        if fileManager.fileExists(atPath: legacy.path) {
-            try fileManager.removeItem(at: legacy)
-        }
-        let directory = runDirectory.appendingPathComponent(directoryName, isDirectory: true)
-        if fileManager.fileExists(atPath: directory.path) {
-            try fileManager.removeItem(at: directory)
-        }
-    }
-
-    private static func readProcessID(at url: URL, fileManager: FileManager) -> Int32? {
-        guard fileManager.fileExists(atPath: url.path),
-              let raw = try? String(contentsOf: url, encoding: .utf8)
-                .trimmingCharacters(in: .whitespacesAndNewlines),
-              let processID = Int32(raw),
-              processID > 1 else {
-            return nil
-        }
-        return processID
-    }
-}
-
-private final class WorkflowProcessStderrTail: @unchecked Sendable {
-    private static let maximumBytes = 16 * 1_024
-    private let lock = NSLock()
-    private var data = Data()
-
-    func append(_ newData: Data) {
-        lock.lock()
-        defer { lock.unlock() }
-        if newData.count >= Self.maximumBytes {
-            data = Data(newData.suffix(Self.maximumBytes))
-            return
-        }
-        data.append(newData)
-        if data.count > Self.maximumBytes {
-            data.removeFirst(data.count - Self.maximumBytes)
-        }
-    }
-
-    var string: String {
-        lock.lock()
-        defer { lock.unlock() }
-        return String(decoding: data, as: UTF8.self)
-            .trimmingCharacters(in: .whitespacesAndNewlines)
-    }
-}
-
 private struct WorkflowPreparedParallelNode: @unchecked Sendable {
     let node: WorkflowNode
     let index: Int
@@ -274,186 +65,6 @@ private final class WorkflowParallelOutcomeBox: @unchecked Sendable {
     }
 }
 
-protocol WorkflowProcessRunning {
-    func run(arguments: [String], currentDirectory: URL) throws -> WorkflowProcessResult
-}
-
-protocol WorkflowStreamingProcessRunning {
-    func run(
-        executable: URL,
-        arguments: [String],
-        currentDirectory: URL,
-        timeoutSeconds: Int?,
-        stdoutLineHandler: ((String) throws -> Void)?
-    ) throws -> WorkflowProcessResult
-
-    func run(
-        executable: URL,
-        arguments: [String],
-        currentDirectory: URL,
-        timeoutSeconds: Int?,
-        stdoutLineHandler: ((String) throws -> Void)?,
-        stdoutChunkHandler: ((Data) throws -> Void)?
-    ) throws -> WorkflowProcessResult
-}
-
-extension WorkflowStreamingProcessRunning {
-    // Chunk delivery is opt-in for runners that support it; others keep
-    // line-based behavior and ignore raw chunks.
-    func run(
-        executable: URL,
-        arguments: [String],
-        currentDirectory: URL,
-        timeoutSeconds: Int?,
-        stdoutLineHandler: ((String) throws -> Void)?,
-        stdoutChunkHandler: ((Data) throws -> Void)?
-    ) throws -> WorkflowProcessResult {
-        try run(
-            executable: executable,
-            arguments: arguments,
-            currentDirectory: currentDirectory,
-            timeoutSeconds: timeoutSeconds,
-            stdoutLineHandler: stdoutLineHandler
-        )
-    }
-}
-
-struct WorkflowProcessRunner: WorkflowProcessRunning, WorkflowStreamingProcessRunning {
-    static func stdoutCaptureURL(in directory: URL) -> URL {
-        directory.appendingPathComponent(".workflow-stdout-\(UUID().uuidString)")
-    }
-
-    func run(arguments: [String], currentDirectory: URL) throws -> WorkflowProcessResult {
-        try run(
-            executable: CurrentExecutable.url(),
-            arguments: arguments,
-            currentDirectory: currentDirectory,
-            timeoutSeconds: nil,
-            stdoutLineHandler: nil
-        )
-    }
-
-    func run(
-        executable: URL,
-        arguments: [String],
-        currentDirectory: URL,
-        timeoutSeconds: Int?,
-        stdoutLineHandler: ((String) throws -> Void)?
-    ) throws -> WorkflowProcessResult {
-        try run(
-            executable: executable,
-            arguments: arguments,
-            currentDirectory: currentDirectory,
-            timeoutSeconds: timeoutSeconds,
-            stdoutLineHandler: stdoutLineHandler,
-            stdoutChunkHandler: nil
-        )
-    }
-
-    func run(
-        executable: URL,
-        arguments: [String],
-        currentDirectory: URL,
-        timeoutSeconds: Int?,
-        stdoutLineHandler: ((String) throws -> Void)?,
-        stdoutChunkHandler: ((Data) throws -> Void)?
-    ) throws -> WorkflowProcessResult {
-        let process = Process()
-        process.executableURL = executable
-        process.arguments = arguments
-        process.currentDirectoryURL = currentDirectory
-        let stdout = Pipe()
-        let stderr = Pipe()
-        let stderrTail = WorkflowProcessStderrTail()
-        let runDirectory = currentDirectory.deletingLastPathComponent().deletingLastPathComponent()
-        process.standardInput = FileHandle.nullDevice
-        process.standardOutput = stdout
-        process.standardError = stderr
-        stderr.fileHandleForReading.readabilityHandler = { handle in
-            let data = handle.availableData
-            guard !data.isEmpty else { return }
-            stderrTail.append(data)
-            try? FileHandle.standardError.write(contentsOf: data)
-        }
-        try process.run()
-        do {
-            try WorkflowChildProcessRegistry.register(process.processIdentifier, in: runDirectory)
-        } catch {
-            process.terminate()
-            process.waitUntilExit()
-            throw error
-        }
-        defer {
-            WorkflowChildProcessRegistry.unregister(process.processIdentifier, in: runDirectory)
-        }
-        if FileManager.default.fileExists(atPath: runDirectory.appendingPathComponent("cancel.request").path) {
-            process.terminate()
-        }
-        let timeoutState = WorkflowProcessTimeoutState()
-        let timeoutWorkItem: DispatchWorkItem?
-        if let timeoutSeconds {
-            let workItem = DispatchWorkItem {
-                guard process.isRunning else { return }
-                timeoutState.markTimedOut()
-                process.terminate()
-                Thread.sleep(forTimeInterval: 2)
-                if process.isRunning {
-                    kill(process.processIdentifier, SIGKILL)
-                }
-            }
-            timeoutWorkItem = workItem
-            DispatchQueue.global(qos: .utility).asyncAfter(
-                deadline: .now() + .seconds(timeoutSeconds),
-                execute: workItem
-            )
-        } else {
-            timeoutWorkItem = nil
-        }
-        defer { timeoutWorkItem?.cancel() }
-        var captured = Data()
-        var pending = Data()
-        while true {
-            let data = stdout.fileHandleForReading.availableData
-            if data.isEmpty { break }
-            captured.append(data)
-            try stdoutChunkHandler?(data)
-            guard stdoutLineHandler != nil else { continue }
-            pending.append(data)
-            while let newline = pending.firstIndex(of: 0x0A) {
-                let lineData = pending.prefix(upTo: newline)
-                pending.removeSubrange(...newline)
-                let line = String(decoding: lineData, as: UTF8.self)
-                    .trimmingCharacters(in: .whitespacesAndNewlines)
-                if !line.isEmpty { try stdoutLineHandler?(line) }
-            }
-        }
-        process.waitUntilExit()
-        stderr.fileHandleForReading.readabilityHandler = nil
-        let remainingStderr = stderr.fileHandleForReading.readDataToEndOfFile()
-        if !remainingStderr.isEmpty {
-            stderrTail.append(remainingStderr)
-            try? FileHandle.standardError.write(contentsOf: remainingStderr)
-        }
-        if !pending.isEmpty {
-            let line = String(decoding: pending, as: UTF8.self)
-                .trimmingCharacters(in: .whitespacesAndNewlines)
-            if !line.isEmpty { try stdoutLineHandler?(line) }
-        }
-        if FileManager.default.fileExists(atPath: runDirectory.appendingPathComponent("cancel.request").path) {
-            throw WorkflowCancellationError()
-        }
-        if timeoutState.didTimeOut, let timeoutSeconds {
-            throw WorkflowProcessTimeoutError(seconds: timeoutSeconds)
-        }
-        return WorkflowProcessResult(
-            status: process.terminationStatus,
-            stdout: String(decoding: captured, as: UTF8.self).trimmingCharacters(in: .whitespacesAndNewlines),
-            stderr: stderrTail.string,
-            terminationReason: WorkflowProcessTerminationReason(process.terminationReason)
-        )
-    }
-}
-
 struct WorkflowRunner: @unchecked Sendable {
     let bundleDirectory: URL
     let runDirectory: URL
@@ -464,6 +75,8 @@ struct WorkflowRunner: @unchecked Sendable {
     let cacheDirectory: URL?
     let now: () -> Date
     let eventHandler: ((GraphRunEvent) -> Void)?
+    let artifactStore: WorkflowArtifactStore
+    let runStore: WorkflowRunStore
 
     init(
         bundleDirectory: URL,
@@ -489,6 +102,14 @@ struct WorkflowRunner: @unchecked Sendable {
                 : nil)
         self.now = now
         self.eventHandler = eventHandler
+        self.artifactStore = WorkflowArtifactStore(
+            bundleDirectory: self.bundleDirectory, runDirectory: self.runDirectory,
+            cacheDirectory: self.cacheDirectory, resume: resume, fileManager: fileManager
+        )
+        self.runStore = WorkflowRunStore(
+            bundleDirectory: self.bundleDirectory, runDirectory: self.runDirectory,
+            resume: resume, executor: executor, fileManager: fileManager, now: now, eventHandler: eventHandler
+        )
     }
 
     func execute() throws -> WorkflowRunOutcome {
@@ -541,12 +162,13 @@ struct WorkflowRunner: @unchecked Sendable {
             throw ValidationError(validation.diagnostics.map(\.message).joined(separator: " "))
         }
 
-        try prepareRunDirectory()
-        let localizedInputs = try localizeInputs(graph: graph, inputs: inputs, assets: assets)
-        var manifest = try initialManifest(graph: graph, job: job, order: validation.order)
-        var sequence = try existingEventCount()
-        try persist(manifest)
-        try record(
+        let session = try runStore.prepare(graph: graph, job: job, order: validation.order)
+        defer { session.lease.release() }
+        let localizedInputs = try artifactStore.localizeInputs(graph: graph, inputs: inputs, assets: assets)
+        var manifest = session.manifest
+        var sequence = session.eventCount
+        try runStore.persist(manifest)
+        try runStore.record(
             GraphRunEvent(
                 sequence: sequence,
                 createdAt: now(),
@@ -559,7 +181,7 @@ struct WorkflowRunner: @unchecked Sendable {
         sequence += 1
         manifest.state = .running
         manifest.updatedAt = now()
-        try persist(manifest)
+        try runStore.persist(manifest)
 
         var nodeOutputs: [String: [String: WorkflowValue]] = [:]
         do {
@@ -579,7 +201,7 @@ struct WorkflowRunner: @unchecked Sendable {
                       let nodeIndex = manifest.nodes.firstIndex(where: { $0.id == nodeID }) else {
                     throw ValidationError("Workflow execution order referenced missing node '\(nodeID)'.")
                 }
-                if fileManager.fileExists(atPath: runDirectory.appendingPathComponent("cancel.request").path) {
+                if Task.isCancelled || fileManager.fileExists(atPath: runDirectory.appendingPathComponent("cancel.request").path) {
                     throw WorkflowCancellationError()
                 }
 
@@ -613,12 +235,12 @@ struct WorkflowRunner: @unchecked Sendable {
                     models: nodeModels,
                     upstreamOutputs: upstreamOutputs
                 ))
-                if workflowNodeAllowsResumeReuse(node), try shouldResume(
+                if workflowNodeAllowsResumeReuse(node), try artifactStore.shouldResume(
                     manifest.nodes[nodeIndex],
                     expectedFingerprint: fingerprint,
                     nodeOutputs: &nodeOutputs
                 ) {
-                    try record(.init(
+                    try runStore.record(.init(
                         sequence: sequence,
                         createdAt: now(),
                         type: "node_resumed",
@@ -640,7 +262,7 @@ struct WorkflowRunner: @unchecked Sendable {
                 manifest.nodes[nodeIndex].models = nodeModels
                 manifest.nodes[nodeIndex].maxAttempts = node.execution?.resolvedMaxAttempts ?? 1
                 if (node.execution?.resolvedCache ?? .automatic) == .automatic,
-                   let cached = try restoreCachedOutputs(
+                   let cached = try artifactStore.restoreCachedOutputs(
                     fingerprint: fingerprint,
                     invocation: invocation,
                     node: node,
@@ -654,8 +276,8 @@ struct WorkflowRunner: @unchecked Sendable {
                     manifest.nodes[nodeIndex].startedAt = now()
                     manifest.nodes[nodeIndex].completedAt = now()
                     manifest.updatedAt = now()
-                    try persist(manifest)
-                    try record(.init(
+                    try runStore.persist(manifest)
+                    try runStore.record(.init(
                         sequence: sequence,
                         createdAt: now(),
                         type: "node_cache_hit",
@@ -669,8 +291,8 @@ struct WorkflowRunner: @unchecked Sendable {
                 manifest.nodes[nodeIndex].state = .preflighting
                 manifest.nodes[nodeIndex].startedAt = now()
                 manifest.updatedAt = now()
-                try persist(manifest)
-                try record(.init(
+                try runStore.persist(manifest)
+                try runStore.record(.init(
                     sequence: sequence,
                     createdAt: now(),
                     type: "node_preflight_started",
@@ -717,8 +339,8 @@ struct WorkflowRunner: @unchecked Sendable {
                     manifest.nodes[nodeIndex].exitStatus = nil
                     manifest.nodes[nodeIndex].error = nil
                     manifest.updatedAt = now()
-                    try persist(manifest)
-                    try record(.init(
+                    try runStore.persist(manifest)
+                    try runStore.record(.init(
                         sequence: sequence,
                         createdAt: now(),
                         type: "node_started",
@@ -752,19 +374,19 @@ struct WorkflowRunner: @unchecked Sendable {
                                 } else {
                                     let runArtifact: GraphRunEventArtifact?
                                     if let eventArtifact = event.artifact {
-                                        let artifactURL = try invocationOutputURL(
+                                        let artifactURL = try artifactStore.invocationOutputURL(
                                             eventArtifact.path,
                                             nodeDirectory: nodeDirectory
                                         )
                                         runArtifact = GraphRunEventArtifact(
                                             name: eventArtifact.name,
-                                            path: try portableArtifactPath(for: artifactURL),
+                                            path: try artifactStore.portableArtifactPath(for: artifactURL),
                                             contentType: eventArtifact.contentType
                                         )
                                     } else {
                                         runArtifact = nil
                                     }
-                                    try record(event.runEvent(
+                                    try runStore.record(event.runEvent(
                                         sequence: sequence,
                                         nodeID: nodeID,
                                         artifact: runArtifact
@@ -780,7 +402,7 @@ struct WorkflowRunner: @unchecked Sendable {
                                 let stamp = now()
                                 guard stamp.timeIntervalSince(lastDeltaAt) >= 0.3 else { return }
                                 lastDeltaAt = stamp
-                                try record(.init(
+                                try runStore.record(.init(
                                     sequence: sequence,
                                     createdAt: stamp,
                                     type: "node_output_delta",
@@ -804,7 +426,7 @@ struct WorkflowRunner: @unchecked Sendable {
                         guard result.status == 0 else {
                             throw ValidationError("Node '\(nodeID)' \(result.failureSummary).")
                         }
-                        verifiedOutputs = try verifyOutputs(
+                        verifiedOutputs = try artifactStore.verifyOutputs(
                             invocation.outputs,
                             providerValues: providerOutputs,
                             node: node,
@@ -816,9 +438,9 @@ struct WorkflowRunner: @unchecked Sendable {
                         let message = (error as? ValidationError)?.message ?? error.localizedDescription
                         manifest.nodes[nodeIndex].error = message
                         manifest.updatedAt = now()
-                        try persist(manifest)
+                        try runStore.persist(manifest)
                         guard attempt < maxAttempts else { throw error }
-                        try record(.init(
+                        try runStore.record(.init(
                             sequence: sequence,
                             createdAt: now(),
                             type: "node_retrying",
@@ -827,7 +449,7 @@ struct WorkflowRunner: @unchecked Sendable {
                             message: "Attempt \(attempt) failed: \(message)"
                         ))
                         sequence += 1
-                        try clearAttemptOutputs(invocation.outputs, nodeDirectory: nodeDirectory)
+                        try artifactStore.clearAttemptOutputs(invocation.outputs, nodeDirectory: nodeDirectory)
                     }
                 }
 
@@ -839,14 +461,14 @@ struct WorkflowRunner: @unchecked Sendable {
                 manifest.nodes[nodeIndex].outputs = verified.outputs
                 if node.execution?.resolvedCache != .never {
                     do {
-                        try storeCachedOutputs(
+                        try artifactStore.storeCachedOutputs(
                             verified,
                             fingerprint: fingerprint,
                             policy: node.execution?.resolvedCache ?? .automatic,
                             nodeDirectory: nodeDirectory
                         )
                         if cacheDirectory != nil {
-                            try record(.init(
+                            try runStore.record(.init(
                                 sequence: sequence,
                                 createdAt: now(),
                                 type: "node_cache_stored",
@@ -857,7 +479,7 @@ struct WorkflowRunner: @unchecked Sendable {
                             sequence += 1
                         }
                     } catch {
-                        try record(.init(
+                        try runStore.record(.init(
                             sequence: sequence,
                             createdAt: now(),
                             type: "node_cache_store_failed",
@@ -871,8 +493,8 @@ struct WorkflowRunner: @unchecked Sendable {
                 manifest.nodes[nodeIndex].state = .finished
                 manifest.nodes[nodeIndex].completedAt = now()
                 manifest.updatedAt = now()
-                try persist(manifest)
-                try record(.init(
+                try runStore.persist(manifest)
+                try runStore.record(.init(
                     sequence: sequence,
                     createdAt: now(),
                     type: "node_finished",
@@ -884,11 +506,11 @@ struct WorkflowRunner: @unchecked Sendable {
             }
             }
 
-            manifest.outputs = try materializeGraphOutputs(graph: graph, nodeOutputs: nodeOutputs)
+            manifest.outputs = try artifactStore.materializeGraphOutputs(graph: graph, nodeOutputs: nodeOutputs)
             manifest.state = .finished
             manifest.updatedAt = now()
-            try persist(manifest)
-            try record(.init(
+            try runStore.persist(manifest)
+            try runStore.record(.init(
                 sequence: sequence,
                 createdAt: now(),
                 type: "run_finished",
@@ -900,8 +522,8 @@ struct WorkflowRunner: @unchecked Sendable {
             manifest.state = .cancelled
             manifest.error = "Workflow cancellation requested."
             manifest.updatedAt = now()
-            try persist(manifest)
-            try record(.init(
+            try runStore.persist(manifest)
+            try runStore.record(.init(
                 sequence: sequence,
                 createdAt: now(),
                 type: "run_cancelled",
@@ -919,8 +541,8 @@ struct WorkflowRunner: @unchecked Sendable {
             manifest.state = .failed
             manifest.error = message
             manifest.updatedAt = now()
-            try persist(manifest)
-            try record(.init(
+            try runStore.persist(manifest)
+            try runStore.record(.init(
                 sequence: sequence,
                 createdAt: now(),
                 type: "run_failed",
@@ -998,12 +620,12 @@ struct WorkflowRunner: @unchecked Sendable {
                     models: models,
                     upstreamOutputs: upstreamOutputs
                 ))
-                if workflowNodeAllowsResumeReuse(node), try shouldResume(
+                if workflowNodeAllowsResumeReuse(node), try artifactStore.shouldResume(
                     manifest.nodes[nodeIndex],
                     expectedFingerprint: fingerprint,
                     nodeOutputs: &nodeOutputs
                 ) {
-                    try record(.init(
+                    try runStore.record(.init(
                         sequence: sequence,
                         createdAt: now(),
                         type: "node_resumed",
@@ -1028,7 +650,7 @@ struct WorkflowRunner: @unchecked Sendable {
                 manifest.nodes[nodeIndex].models = models
                 manifest.nodes[nodeIndex].maxAttempts = node.execution?.resolvedMaxAttempts ?? 1
                 if (node.execution?.resolvedCache ?? .automatic) == .automatic,
-                   let cached = try restoreCachedOutputs(
+                   let cached = try artifactStore.restoreCachedOutputs(
                     fingerprint: fingerprint,
                     invocation: invocation,
                     node: node,
@@ -1042,8 +664,8 @@ struct WorkflowRunner: @unchecked Sendable {
                     manifest.nodes[nodeIndex].startedAt = now()
                     manifest.nodes[nodeIndex].completedAt = now()
                     manifest.updatedAt = now()
-                    try persist(manifest)
-                    try record(.init(
+                    try runStore.persist(manifest)
+                    try runStore.record(.init(
                         sequence: sequence,
                         createdAt: now(),
                         type: "node_cache_hit",
@@ -1060,8 +682,8 @@ struct WorkflowRunner: @unchecked Sendable {
                 manifest.nodes[nodeIndex].state = .preflighting
                 manifest.nodes[nodeIndex].startedAt = now()
                 manifest.updatedAt = now()
-                try persist(manifest)
-                try record(.init(
+                try runStore.persist(manifest)
+                try runStore.record(.init(
                     sequence: sequence,
                     createdAt: now(),
                     type: "node_preflight_started",
@@ -1100,8 +722,8 @@ struct WorkflowRunner: @unchecked Sendable {
                 manifest.nodes[nodeIndex].error = nil
                 manifest.nodes[nodeIndex].exitStatus = nil
                 manifest.updatedAt = now()
-                try persist(manifest)
-                try record(.init(
+                try runStore.persist(manifest)
+                try runStore.record(.init(
                     sequence: sequence,
                     createdAt: now(),
                     type: "node_started",
@@ -1147,25 +769,25 @@ struct WorkflowRunner: @unchecked Sendable {
                     case .provider(let providerEvent):
                         let artifact: GraphRunEventArtifact?
                         if let eventArtifact = providerEvent.artifact {
-                            let artifactURL = try invocationOutputURL(
+                            let artifactURL = try artifactStore.invocationOutputURL(
                                 eventArtifact.path,
                                 nodeDirectory: item.directory
                             )
                             artifact = GraphRunEventArtifact(
                                 name: eventArtifact.name,
-                                path: try portableArtifactPath(for: artifactURL),
+                                path: try artifactStore.portableArtifactPath(for: artifactURL),
                                 contentType: eventArtifact.contentType
                             )
                         } else {
                             artifact = nil
                         }
-                        try record(providerEvent.runEvent(
+                        try runStore.record(providerEvent.runEvent(
                             sequence: sequence,
                             nodeID: item.node.id,
                             artifact: artifact
                         ))
                     case .retrying(let attempt, let message):
-                        try record(.init(
+                        try runStore.record(.init(
                             sequence: sequence,
                             createdAt: now(),
                             type: "node_retrying",
@@ -1174,7 +796,7 @@ struct WorkflowRunner: @unchecked Sendable {
                             message: "Attempt \(attempt) failed: \(message)"
                         ))
                     case .started(let attempt):
-                        try record(.init(
+                        try runStore.record(.init(
                             sequence: sequence,
                             createdAt: now(),
                             type: "node_started",
@@ -1200,14 +822,14 @@ struct WorkflowRunner: @unchecked Sendable {
                 manifest.nodes[item.index].outputs = verified.outputs
                 if item.node.execution?.resolvedCache != .never {
                     do {
-                        try storeCachedOutputs(
+                        try artifactStore.storeCachedOutputs(
                             verified,
                             fingerprint: item.fingerprint,
                             policy: item.node.execution?.resolvedCache ?? .automatic,
                             nodeDirectory: item.directory
                         )
                         if cacheDirectory != nil {
-                            try record(.init(
+                            try runStore.record(.init(
                                 sequence: sequence,
                                 createdAt: now(),
                                 type: "node_cache_stored",
@@ -1218,7 +840,7 @@ struct WorkflowRunner: @unchecked Sendable {
                             sequence += 1
                         }
                     } catch {
-                        try record(.init(
+                        try runStore.record(.init(
                             sequence: sequence,
                             createdAt: now(),
                             type: "node_cache_store_failed",
@@ -1234,7 +856,7 @@ struct WorkflowRunner: @unchecked Sendable {
                 manifest.nodes[item.index].error = nil
                 completed.insert(item.node.id)
                 pending.removeAll { $0 == item.node.id }
-                try record(.init(
+                try runStore.record(.init(
                     sequence: sequence,
                     createdAt: now(),
                     type: "node_finished",
@@ -1245,7 +867,7 @@ struct WorkflowRunner: @unchecked Sendable {
                 sequence += 1
             }
             manifest.updatedAt = now()
-            try persist(manifest)
+            try runStore.persist(manifest)
             if wasCancelled { throw WorkflowCancellationError() }
             if let firstFailure { throw ValidationError(firstFailure) }
         }
@@ -1298,7 +920,7 @@ struct WorkflowRunner: @unchecked Sendable {
                 guard result.status == 0 else {
                     throw ValidationError("Node '\(prepared.node.id)' \(result.failureSummary).")
                 }
-                let verified = try verifyOutputs(
+                let verified = try artifactStore.verifyOutputs(
                     prepared.invocation.outputs,
                     providerValues: providerOutputs,
                     node: prepared.node,
@@ -1336,7 +958,7 @@ struct WorkflowRunner: @unchecked Sendable {
                 events.append(.retrying(attempt: attempt, message: message))
                 events.append(.started(attempt: attempt + 1))
                 do {
-                    try clearAttemptOutputs(
+                    try artifactStore.clearAttemptOutputs(
                         prepared.invocation.outputs,
                         nodeDirectory: prepared.directory
                     )
@@ -1458,342 +1080,6 @@ struct WorkflowRunner: @unchecked Sendable {
         return result
     }
 
-    private func restoreCachedOutputs(
-        fingerprint: String,
-        invocation: WorkflowNodeInvocation,
-        node: WorkflowNode,
-        nodeDirectory: URL
-    ) throws -> WorkflowVerifiedNodeOutputs? {
-        guard let cacheDirectory else { return nil }
-        let entry = cacheDirectory.appendingPathComponent(fingerprint, isDirectory: true)
-        let manifestURL = entry.appendingPathComponent(WorkflowNodeCacheManifest.filename)
-        guard fileManager.fileExists(atPath: manifestURL.path) else { return nil }
-        do {
-            let manifest = try WorkflowBundleCodec.decoder().decode(
-                WorkflowNodeCacheManifest.self,
-                from: Data(contentsOf: manifestURL)
-            )
-            guard manifest.contractVersion == WorkflowNodeCacheManifest.contractVersion,
-                  manifest.fingerprint == fingerprint else {
-                throw ValidationError("Node cache manifest does not match its fingerprint.")
-            }
-            try clearAttemptOutputs(invocation.outputs, nodeDirectory: nodeDirectory)
-            var providerValues: [String: WorkflowValue] = [:]
-            let filesRoot = entry.appendingPathComponent("files", isDirectory: true)
-            for output in manifest.outputs {
-                guard let descriptor = invocation.outputs[output.name], descriptor.type == output.type else {
-                    throw ValidationError("Node cache output contract has changed for '\(output.name)'.")
-                }
-                if let relativePath = output.relativePath {
-                    guard let descriptorPath = descriptor.path else {
-                        throw ValidationError("Node cache output '\(output.name)' no longer has a path.")
-                    }
-                    let destination = try invocationOutputURL(descriptorPath, nodeDirectory: nodeDirectory)
-                    guard try nodeRelativePath(destination, nodeDirectory: nodeDirectory) == relativePath else {
-                        throw ValidationError("Node cache output path has changed for '\(output.name)'.")
-                    }
-                    let source = try confinedCacheURL(relativePath, root: filesRoot)
-                    try copyCacheItem(from: source, to: destination)
-                } else if let value = output.value {
-                    providerValues[output.name] = value
-                }
-            }
-            let verified = try verifyOutputs(
-                invocation.outputs,
-                providerValues: providerValues,
-                node: node,
-                nodeDirectory: nodeDirectory
-            )
-            let restoredRecords = try verified.outputs.map {
-                try cacheOutput($0, nodeDirectory: nodeDirectory)
-            }.sorted { $0.name < $1.name }
-            guard restoredRecords == manifest.outputs.sorted(by: { $0.name < $1.name }) else {
-                throw ValidationError("Node cache output hashes do not match its manifest.")
-            }
-            return verified
-        } catch {
-            try? clearAttemptOutputs(invocation.outputs, nodeDirectory: nodeDirectory)
-            try? fileManager.removeItem(at: entry)
-            return nil
-        }
-    }
-
-    private func storeCachedOutputs(
-        _ verified: WorkflowVerifiedNodeOutputs,
-        fingerprint: String,
-        policy: WorkflowNodeCachePolicy,
-        nodeDirectory: URL
-    ) throws {
-        guard let cacheDirectory else { return }
-        try fileManager.createDirectory(at: cacheDirectory, withIntermediateDirectories: true)
-        let entry = cacheDirectory.appendingPathComponent(fingerprint, isDirectory: true)
-        if fileManager.fileExists(atPath: entry.path), policy == .automatic { return }
-        let staging = cacheDirectory.appendingPathComponent(".\(fingerprint).\(UUID().uuidString)", isDirectory: true)
-        defer { try? fileManager.removeItem(at: staging) }
-        let filesRoot = staging.appendingPathComponent("files", isDirectory: true)
-        try fileManager.createDirectory(at: filesRoot, withIntermediateDirectories: true)
-        let outputs = try verified.outputs.map { output -> WorkflowNodeCacheOutput in
-            let record = try cacheOutput(output, nodeDirectory: nodeDirectory)
-            if let relativePath = record.relativePath {
-                let source = try invocationOutputURL(relativePath, nodeDirectory: nodeDirectory)
-                let destination = try confinedCacheURL(relativePath, root: filesRoot)
-                try copyCacheItem(from: source, to: destination)
-            }
-            return record
-        }.sorted { $0.name < $1.name }
-        try WorkflowBundleCodec.write(
-            WorkflowNodeCacheManifest(
-                contractVersion: WorkflowNodeCacheManifest.contractVersion,
-                fingerprint: fingerprint,
-                outputs: outputs
-            ),
-            to: staging.appendingPathComponent(WorkflowNodeCacheManifest.filename)
-        )
-        if fileManager.fileExists(atPath: entry.path) {
-            try fileManager.removeItem(at: entry)
-        }
-        try fileManager.moveItem(at: staging, to: entry)
-    }
-
-    private func cacheOutput(
-        _ output: GraphRunNodeOutput,
-        nodeDirectory: URL
-    ) throws -> WorkflowNodeCacheOutput {
-        guard let sha256 = output.sha256 else {
-            throw ValidationError("Workflow output '\(output.name)' has no cache fingerprint.")
-        }
-        let relativePath: String?
-        if let path = output.path {
-            relativePath = try nodeRelativePath(artifactURL(for: path), nodeDirectory: nodeDirectory)
-        } else {
-            relativePath = nil
-        }
-        return WorkflowNodeCacheOutput(
-            name: output.name,
-            type: output.type,
-            value: output.value,
-            relativePath: relativePath,
-            contentType: output.contentType,
-            sizeBytes: output.sizeBytes,
-            sha256: sha256
-        )
-    }
-
-    private func nodeRelativePath(_ url: URL, nodeDirectory: URL) throws -> String {
-        let candidate = url.standardizedFileURL.path
-        let root = nodeDirectory.standardizedFileURL.path
-        guard candidate.hasPrefix(root + "/") else {
-            throw ValidationError("Workflow cache output escapes the node directory: \(candidate)")
-        }
-        return String(candidate.dropFirst(root.count + 1))
-    }
-
-    private func confinedCacheURL(_ path: String, root: URL) throws -> URL {
-        guard isConfinedRelativeWorkflowPath(path) else {
-            throw ValidationError("Workflow cache contains an unconfined path: \(path)")
-        }
-        let candidate = root.appendingPathComponent(path).standardizedFileURL
-        guard candidate.path.hasPrefix(root.standardizedFileURL.path + "/") else {
-            throw ValidationError("Workflow cache path escapes its entry: \(path)")
-        }
-        return candidate
-    }
-
-    private func copyCacheItem(from source: URL, to destination: URL) throws {
-        let values = try source.resourceValues(forKeys: [.isSymbolicLinkKey])
-        guard values.isSymbolicLink != true else {
-            throw ValidationError("Workflow cache items cannot be symbolic links: \(source.path)")
-        }
-        if fileManager.fileExists(atPath: destination.path) {
-            try fileManager.removeItem(at: destination)
-        }
-        try fileManager.createDirectory(
-            at: destination.deletingLastPathComponent(),
-            withIntermediateDirectories: true
-        )
-        try fileManager.copyItem(at: source, to: destination)
-    }
-
-    private func clearAttemptOutputs(
-        _ outputs: [String: WorkflowInvocationOutput],
-        nodeDirectory: URL
-    ) throws {
-        for descriptor in outputs.values {
-            guard let path = descriptor.path else { continue }
-            let outputURL = try invocationOutputURL(path, nodeDirectory: nodeDirectory)
-            if fileManager.fileExists(atPath: outputURL.path) {
-                try fileManager.removeItem(at: outputURL)
-            }
-        }
-        let artifacts = nodeDirectory.appendingPathComponent("artifacts", isDirectory: true)
-        if fileManager.fileExists(atPath: artifacts.path) {
-            try fileManager.removeItem(at: artifacts)
-        }
-        let stdout = nodeDirectory.appendingPathComponent("stdout.txt")
-        if fileManager.fileExists(atPath: stdout.path) {
-            try fileManager.removeItem(at: stdout)
-        }
-    }
-
-    private func verifyOutputs(
-        _ descriptors: [String: WorkflowInvocationOutput],
-        providerValues: [String: WorkflowValue]?,
-        node: WorkflowNode,
-        nodeDirectory: URL
-    ) throws -> WorkflowVerifiedNodeOutputs {
-        var artifacts: [GraphRunArtifact] = []
-        var records: [GraphRunNodeOutput] = []
-        var values: [String: WorkflowValue] = [:]
-        for name in descriptors.keys.sorted() {
-            guard let descriptor = descriptors[name] else { continue }
-            let providerValue = providerValues?[name]
-            switch descriptor.type {
-            case .asset, .assetCollection, .assetArray:
-                guard let path = descriptor.path else {
-                    throw ValidationError("Node '\(node.id)' output '\(name)' has no declared path.")
-                }
-                let url = try invocationOutputURL(path, nodeDirectory: nodeDirectory)
-                if providerValue == .null || !fileManager.fileExists(atPath: url.path) {
-                    if descriptor.optional { continue }
-                    throw ValidationError("Node '\(node.id)' did not produce declared output '\(name)'.")
-                }
-                if let providerPath = providerValue?.stringValue {
-                    let reported = try invocationOutputURL(providerPath, nodeDirectory: nodeDirectory)
-                    guard reported.standardizedFileURL == url.standardizedFileURL else {
-                        throw ValidationError("Node '\(node.id)' reported an unexpected path for output '\(name)'.")
-                    }
-                }
-                let outputArtifact = try artifact(
-                    name: name,
-                    nodeKind: node.kind,
-                    url: url,
-                    contentType: descriptor.contentTypes.first
-                )
-                artifacts.append(outputArtifact)
-                records.append(.init(
-                    name: name,
-                    type: descriptor.type,
-                    value: nil,
-                    path: outputArtifact.path,
-                    contentType: outputArtifact.contentType,
-                    sizeBytes: outputArtifact.sizeBytes,
-                    sha256: outputArtifact.sha256
-                ))
-                values[name] = .string(url.path)
-            case .assetDirectory:
-                guard let path = descriptor.path else {
-                    throw ValidationError("Node '\(node.id)' directory output '\(name)' has no declared path.")
-                }
-                let url = try invocationOutputURL(path, nodeDirectory: nodeDirectory)
-                var isDirectory: ObjCBool = false
-                if providerValue == .null || !fileManager.fileExists(atPath: url.path, isDirectory: &isDirectory) {
-                    if descriptor.optional { continue }
-                    throw ValidationError("Node '\(node.id)' did not produce declared directory output '\(name)'.")
-                }
-                guard isDirectory.boolValue else {
-                    throw ValidationError("Node '\(node.id)' output '\(name)' is not a directory.")
-                }
-                if let providerPath = providerValue?.stringValue {
-                    let reported = try invocationOutputURL(providerPath, nodeDirectory: nodeDirectory)
-                    guard reported.standardizedFileURL == url.standardizedFileURL else {
-                        throw ValidationError("Node '\(node.id)' reported an unexpected path for output '\(name)'.")
-                    }
-                }
-                let directory = try directoryIdentity(url)
-                let manifestURL = nodeDirectory
-                    .appendingPathComponent("artifacts", isDirectory: true)
-                    .appendingPathComponent("\(name).manifest.json")
-                try WorkflowBundleCodec.write(directory.manifest, to: manifestURL)
-                let manifestArtifact = try artifact(
-                    name: "\(name)_manifest",
-                    nodeKind: node.kind,
-                    url: manifestURL,
-                    contentType: "application/json"
-                )
-                artifacts.append(manifestArtifact)
-                records.append(.init(
-                    name: name,
-                    type: .assetDirectory,
-                    value: nil,
-                    path: try portableArtifactPath(for: url),
-                    contentType: descriptor.contentTypes.first,
-                    sizeBytes: directory.sizeBytes,
-                    sha256: directory.sha256
-                ))
-                values[name] = .string(url.path)
-            case .string, .integer, .number, .boolean, .enumeration, .json:
-                guard let providerValue, providerValue != .null else {
-                    if descriptor.optional { continue }
-                    throw ValidationError("Node '\(node.id)' did not report declared value output '\(name)'.")
-                }
-                guard workflowValue(providerValue, matches: descriptor.type) else {
-                    throw ValidationError("Node '\(node.id)' output '\(name)' has the wrong value type.")
-                }
-                records.append(.init(
-                    name: name,
-                    type: descriptor.type,
-                    value: providerValue,
-                    path: nil,
-                    contentType: nil,
-                    sizeBytes: nil,
-                    sha256: try WorkflowBundleCodec.hash(providerValue)
-                ))
-                values[name] = providerValue
-            }
-        }
-        if let providerValues {
-            let undeclared = Set(providerValues.keys).subtracting(descriptors.keys)
-            guard undeclared.isEmpty else {
-                throw ValidationError(
-                    "Node '\(node.id)' reported undeclared outputs: \(undeclared.sorted().joined(separator: ", "))."
-                )
-            }
-        }
-        return WorkflowVerifiedNodeOutputs(artifacts: artifacts, outputs: records, values: values)
-    }
-
-    private func invocationOutputURL(_ path: String, nodeDirectory: URL) throws -> URL {
-        let candidate: URL
-        if path.hasPrefix("/") {
-            candidate = URL(fileURLWithPath: path).standardizedFileURL
-        } else {
-            guard isConfinedRelativeWorkflowPath(path) else {
-                throw ValidationError("Workflow provider output path is not confined: \(path)")
-            }
-            candidate = nodeDirectory.appendingPathComponent(path).standardizedFileURL
-        }
-        let root = nodeDirectory.standardizedFileURL.path
-        guard candidate.path.hasPrefix(root + "/") else {
-            throw ValidationError("Workflow provider output path escapes the node directory: \(path)")
-        }
-        return candidate
-    }
-
-    private func directoryIdentity(_ directory: URL) throws -> WorkflowDirectoryIdentity {
-        let root = directory.resolvingSymlinksInPath()
-        var entries: [WorkflowAssetEntry] = []
-        for path in try fileManager.subpathsOfDirectory(atPath: root.path).sorted() {
-            let url = root.appendingPathComponent(path)
-            let values = try url.resourceValues(forKeys: [.isRegularFileKey, .isSymbolicLinkKey])
-            if values.isSymbolicLink == true {
-                throw ValidationError("Workflow output directories cannot contain symbolic links: \(url.path)")
-            }
-            guard values.isRegularFile == true else { continue }
-            entries.append(WorkflowAssetEntry(
-                path: path,
-                digest: try ModelArtifactPin.fileSHA256(url),
-                sizeBytes: try ModelArtifactPin.fileByteCount(url),
-                contentType: contentType(for: url)
-            ))
-        }
-        let manifest = WorkflowOutputDirectoryManifest(contractVersion: "mere.run/output-directory.v1", entries: entries)
-        return WorkflowDirectoryIdentity(
-            manifest: manifest,
-            sizeBytes: entries.reduce(0) { $0 + $1.sizeBytes },
-            sha256: try WorkflowBundleCodec.hash(manifest)
-        )
-    }
-
     private func modelProvenance(
         for node: WorkflowNode,
         arguments: [String: WorkflowValue],
@@ -1827,153 +1113,6 @@ struct WorkflowRunner: @unchecked Sendable {
                 installManifestSHA256: manifestDigest
             )
         }.sorted { $0.id < $1.id }
-    }
-
-    private func prepareRunDirectory() throws {
-        if fileManager.fileExists(atPath: runDirectory.path) {
-            guard resume || bundleDirectory == runDirectory else {
-                throw ValidationError("Run directory already exists. Pass --resume to reuse it: \(runDirectory.path)")
-            }
-        } else {
-            try fileManager.createDirectory(at: runDirectory, withIntermediateDirectories: true)
-        }
-        try fileManager.createDirectory(
-            at: runDirectory.appendingPathComponent("nodes", isDirectory: true),
-            withIntermediateDirectories: true
-        )
-        let cancellationURL = runDirectory.appendingPathComponent("cancel.request")
-        if fileManager.fileExists(atPath: cancellationURL.path) {
-            try fileManager.removeItem(at: cancellationURL)
-        }
-        try WorkflowChildProcessRegistry.clear(in: runDirectory, fileManager: fileManager)
-        try fileManager.createDirectory(
-            at: runDirectory.appendingPathComponent("outputs", isDirectory: true),
-            withIntermediateDirectories: true
-        )
-        for filename in ["graph.json", "inputs.json", WorkflowAssetManifest.filename, WorkflowJobManifest.filename] {
-            let source = bundleDirectory.appendingPathComponent(filename)
-            let destination = runDirectory.appendingPathComponent(filename)
-            if source != destination, !fileManager.fileExists(atPath: destination.path) {
-                try fileManager.copyItem(at: source, to: destination)
-            }
-        }
-        let actionsURL = runDirectory.appendingPathComponent("actions.json")
-        if !fileManager.fileExists(atPath: actionsURL.path) {
-            try WorkflowBundleCodec.write([DeclarativeAction](), to: actionsURL)
-        }
-    }
-
-    private func localizeInputs(
-        graph: WorkflowGraphDocument,
-        inputs: WorkflowInputsDocument,
-        assets: WorkflowAssetManifest
-    ) throws -> WorkflowInputsDocument {
-        var localized = inputs.values
-        let root = runDirectory.appendingPathComponent("inputs", isDirectory: true)
-        try fileManager.createDirectory(at: root, withIntermediateDirectories: true)
-        var groupNames = Set<String>()
-        for group in assets.groups {
-            guard groupNames.insert(group.name).inserted,
-                  group.name.range(of: "^[a-z][a-z0-9-]{0,63}$", options: .regularExpression) != nil else {
-                throw ValidationError("Asset manifest contains an invalid or duplicate group '\(group.name)'.")
-            }
-            let groupRoot = root.appendingPathComponent(group.name, isDirectory: true)
-            if group.kind == .assetDirectory {
-                try fileManager.createDirectory(at: groupRoot, withIntermediateDirectories: true)
-            }
-            for entry in group.entries {
-                guard isConfinedRelativeWorkflowPath(entry.path),
-                      entry.digest.count == 64,
-                      entry.digest.allSatisfy({ $0.isHexDigit && !$0.isUppercase }) else {
-                    throw ValidationError("Workflow asset manifest contains an invalid path or digest.")
-                }
-                let source = bundleDirectory
-                    .appendingPathComponent("assets", isDirectory: true)
-                    .appendingPathComponent("sha256", isDirectory: true)
-                    .appendingPathComponent(entry.digest)
-                guard fileManager.fileExists(atPath: source.path),
-                      try ModelArtifactPin.fileByteCount(source) == entry.sizeBytes,
-                      try ModelArtifactPin.fileSHA256(source) == entry.digest else {
-                    throw ValidationError("Workflow asset digest verification failed for '\(group.name)/\(entry.path)'.")
-                }
-                let destination = group.kind == .asset
-                    ? root.appendingPathComponent("\(group.name)-\(entry.path)")
-                    : groupRoot.appendingPathComponent(entry.path)
-                try fileManager.createDirectory(at: destination.deletingLastPathComponent(), withIntermediateDirectories: true)
-                if !fileManager.fileExists(atPath: destination.path) {
-                    do {
-                        try fileManager.linkItem(at: source, to: destination)
-                    } catch {
-                        try fileManager.copyItem(at: source, to: destination)
-                    }
-                }
-            }
-            if group.kind == .asset, let entry = group.entries.first {
-                guard group.entries.count == 1 else {
-                    throw ValidationError("Asset group '\(group.name)' must contain exactly one file.")
-                }
-                localized[group.name] = .string(root.appendingPathComponent("\(group.name)-\(entry.path)").path)
-            } else {
-                localized[group.name] = .string(groupRoot.path)
-            }
-        }
-        return WorkflowInputsDocument(values: localized)
-    }
-
-    private func initialManifest(
-        graph: WorkflowGraphDocument,
-        job: WorkflowJobManifest,
-        order: [String]
-    ) throws -> GraphRunManifest {
-        let manifestURL = runDirectory.appendingPathComponent(GraphRunManifest.filename)
-        if resume, fileManager.fileExists(atPath: manifestURL.path) {
-            var existing = try WorkflowBundleCodec.decoder().decode(
-                GraphRunManifest.self,
-                from: Data(contentsOf: manifestURL)
-            )
-            guard existing.graphFingerprint == job.graphFingerprint else {
-                throw ValidationError("Cannot resume: workflow graph fingerprint changed.")
-            }
-            existing.attempt += 1
-            existing.state = .planned
-            existing.error = nil
-            existing.executor = executor
-            existing.updatedAt = now()
-            return existing
-        }
-        let nodesByID = Dictionary(uniqueKeysWithValues: graph.nodes.map { ($0.id, $0) })
-        return GraphRunManifest(
-            contractVersion: GraphRunManifest.contractVersion,
-            jobID: job.jobID,
-            graphName: graph.name,
-            graphFingerprint: job.graphFingerprint,
-            sourceGraphFingerprint: job.sourceGraphFingerprint,
-            sourceInputFingerprint: job.sourceInputFingerprint,
-            state: .planned,
-            createdAt: now(),
-            updatedAt: now(),
-            attempt: 1,
-            executor: executor,
-            nodes: order.compactMap { id in
-                nodesByID[id].map {
-                    GraphRunNodeRecord(
-                        id: id,
-                        kind: $0.kind,
-                        state: .planned,
-                        startedAt: nil,
-                        completedAt: nil,
-                        exitStatus: nil,
-                        attempt: 0,
-                        maxAttempts: $0.execution?.resolvedMaxAttempts ?? 1,
-                        fingerprint: "",
-                        artifacts: [],
-                        error: nil
-                    )
-                }
-            },
-            outputs: [],
-            error: nil
-        )
     }
 
     private func resolve(
@@ -2029,177 +1168,13 @@ struct WorkflowRunner: @unchecked Sendable {
         }
     }
 
-    private func shouldResume(
-        _ node: GraphRunNodeRecord,
-        expectedFingerprint: String,
-        nodeOutputs: inout [String: [String: WorkflowValue]]
-    ) throws -> Bool {
-        guard resume,
-              node.state == .finished,
-              node.fingerprint == expectedFingerprint,
-              !node.outputs.isEmpty || !node.artifacts.isEmpty else { return false }
-        var outputs: [String: WorkflowValue] = [:]
-        if !node.outputs.isEmpty {
-            for output in node.outputs {
-                if let value = output.value {
-                    guard try WorkflowBundleCodec.hash(value) == output.sha256 else { return false }
-                    outputs[output.name] = value
-                    continue
-                }
-                guard let path = output.path else { return false }
-                let url = try artifactURL(for: path)
-                if output.type == .assetDirectory {
-                    var isDirectory: ObjCBool = false
-                    guard fileManager.fileExists(atPath: url.path, isDirectory: &isDirectory),
-                          isDirectory.boolValue,
-                          try directoryIdentity(url).sha256 == output.sha256 else { return false }
-                } else {
-                    guard fileManager.fileExists(atPath: url.path),
-                          try ModelArtifactPin.fileByteCount(url) == output.sizeBytes,
-                          try ModelArtifactPin.fileSHA256(url) == output.sha256 else { return false }
-                }
-                outputs[output.name] = .string(url.path)
-            }
-        } else {
-            for artifact in node.artifacts {
-                let url = try artifactURL(for: artifact.path)
-                guard fileManager.fileExists(atPath: url.path),
-                      try ModelArtifactPin.fileByteCount(url) == artifact.sizeBytes,
-                      try ModelArtifactPin.fileSHA256(url) == artifact.sha256 else {
-                    return false
-                }
-                outputs[artifact.name] = .string(url.path)
-            }
-        }
-        nodeOutputs[node.id] = outputs
-        return true
-    }
-
-    private func materializeGraphOutputs(
-        graph: WorkflowGraphDocument,
-        nodeOutputs: [String: [String: WorkflowValue]]
-    ) throws -> [GraphRunArtifact] {
-        var artifacts: [GraphRunArtifact] = []
-        let outputsRoot = runDirectory.appendingPathComponent("outputs", isDirectory: true)
-        for name in graph.outputs.keys.sorted() {
-            guard case .reference(let rawReference)? = graph.outputs[name] else { continue }
-            let reference = try WorkflowReference(rawReference)
-            guard case .nodeOutput(let nodeID, let output) = reference.source,
-                  let sourceValue = nodeOutputs[nodeID]?[output],
-                  let node = graph.nodes.first(where: { $0.id == nodeID }),
-                  let outputContract = WorkflowNodeRegistry.output(node: node, name: output) else {
-                throw ValidationError("Workflow output '\(name)' was not produced.")
-            }
-            if outputContract.type != .asset && outputContract.type != .assetCollection && outputContract.type != .assetArray {
-                let destination = outputsRoot.appendingPathComponent("\(name).json")
-                try WorkflowBundleCodec.encoder().encode(sourceValue).write(to: destination, options: .atomic)
-                artifacts.append(try artifact(
-                    name: name,
-                    nodeKind: "graph.output",
-                    url: destination,
-                    contentType: "application/json"
-                ))
-                continue
-            }
-            guard let sourcePath = sourceValue.stringValue else {
-                throw ValidationError("Workflow output '\(name)' did not resolve to an artifact path.")
-            }
-            let source = URL(fileURLWithPath: sourcePath)
-            let suffix = source.pathExtension.isEmpty ? "" : ".\(source.pathExtension)"
-            let destination = outputsRoot.appendingPathComponent("\(name)\(suffix)")
-            if fileManager.fileExists(atPath: destination.path) {
-                try fileManager.removeItem(at: destination)
-            }
-            do {
-                try fileManager.linkItem(at: source, to: destination)
-            } catch {
-                try fileManager.copyItem(at: source, to: destination)
-            }
-            artifacts.append(try artifact(name: name, nodeKind: "graph.output", url: destination))
-        }
-        return artifacts
-    }
-
-    private func artifact(
-        name: String,
-        nodeKind: String,
-        url: URL,
-        contentType explicitContentType: String? = nil
-    ) throws -> GraphRunArtifact {
-        GraphRunArtifact(
-            name: name,
-            kind: nodeKind,
-            path: try portableArtifactPath(for: url),
-            contentType: explicitContentType ?? contentType(for: url),
-            sizeBytes: try ModelArtifactPin.fileByteCount(url),
-            sha256: try ModelArtifactPin.fileSHA256(url)
-        )
-    }
-
     private func throwIfCancellationRequested() throws {
-        if fileManager.fileExists(atPath: runDirectory.appendingPathComponent("cancel.request").path) {
+        if Task.isCancelled || fileManager.fileExists(atPath: runDirectory.appendingPathComponent("cancel.request").path) {
             throw WorkflowCancellationError()
         }
     }
 
-    private func artifactURL(for path: String) throws -> URL {
-        let candidate = path.hasPrefix("/")
-            ? URL(fileURLWithPath: path).standardizedFileURL
-            : runDirectory.appendingPathComponent(path).standardizedFileURL
-        let root = runDirectory.standardizedFileURL.path
-        guard candidate.path == root || candidate.path.hasPrefix(root + "/") else {
-            throw ValidationError("Workflow artifact path escapes the run directory: \(path)")
-        }
-        return candidate
-    }
 
-    private func portableArtifactPath(for url: URL) throws -> String {
-        let candidate = url.standardizedFileURL.path
-        let root = runDirectory.standardizedFileURL.path
-        guard candidate.hasPrefix(root + "/") else {
-            throw ValidationError("Workflow artifact path escapes the run directory: \(candidate)")
-        }
-        return String(candidate.dropFirst(root.count + 1))
-    }
-
-    private func contentType(for url: URL) -> String {
-        switch url.pathExtension.lowercased() {
-        case "png": "image/png"
-        case "jpg", "jpeg": "image/jpeg"
-        case "webp": "image/webp"
-        case "mp4": "video/mp4"
-        case "wav": "audio/wav"
-        case "tif", "tiff": "image/tiff"
-        case "json": "application/json"
-        case "txt": "text/plain"
-        case "safetensors": "application/x-safetensors"
-        default: "application/octet-stream"
-        }
-    }
-
-    private func persist(_ manifest: GraphRunManifest) throws {
-        try WorkflowBundleCodec.write(manifest, to: runDirectory.appendingPathComponent(GraphRunManifest.filename))
-    }
-
-    private func record(_ event: GraphRunEvent) throws {
-        let data = try WorkflowBundleCodec.lineEncoder().encode(event)
-        let url = runDirectory.appendingPathComponent("events.jsonl")
-        if !fileManager.fileExists(atPath: url.path) {
-            try Data().write(to: url)
-        }
-        let handle = try FileHandle(forWritingTo: url)
-        defer { try? handle.close() }
-        try handle.seekToEnd()
-        try handle.write(contentsOf: data)
-        try handle.write(contentsOf: Data("\n".utf8))
-        eventHandler?(event)
-    }
-
-    private func existingEventCount() throws -> Int {
-        let url = runDirectory.appendingPathComponent("events.jsonl")
-        guard fileManager.fileExists(atPath: url.path) else { return 0 }
-        return try String(contentsOf: url, encoding: .utf8).split(separator: "\n").count
-    }
 }
 
 private struct WorkflowNodeFingerprint: Codable {
@@ -2237,63 +1212,6 @@ private struct WorkflowNodeOutputFingerprint: Codable {
         case value
         case sha256
     }
-}
-
-private struct WorkflowVerifiedNodeOutputs {
-    let artifacts: [GraphRunArtifact]
-    let outputs: [GraphRunNodeOutput]
-    let values: [String: WorkflowValue]
-}
-
-private struct WorkflowNodeCacheManifest: Codable {
-    static let contractVersion = "mere.run/node-cache.v1"
-    static let filename = "cache.json"
-
-    let contractVersion: String
-    let fingerprint: String
-    let outputs: [WorkflowNodeCacheOutput]
-
-    enum CodingKeys: String, CodingKey {
-        case contractVersion = "contract_version"
-        case fingerprint
-        case outputs
-    }
-}
-
-private struct WorkflowNodeCacheOutput: Codable, Equatable {
-    let name: String
-    let type: WorkflowPortType
-    let value: WorkflowValue?
-    let relativePath: String?
-    let contentType: String?
-    let sizeBytes: Int64?
-    let sha256: String
-
-    enum CodingKeys: String, CodingKey {
-        case name
-        case type
-        case value
-        case relativePath = "relative_path"
-        case contentType = "content_type"
-        case sizeBytes = "size_bytes"
-        case sha256
-    }
-}
-
-private struct WorkflowOutputDirectoryManifest: Codable {
-    let contractVersion: String
-    let entries: [WorkflowAssetEntry]
-
-    enum CodingKeys: String, CodingKey {
-        case contractVersion = "contract_version"
-        case entries
-    }
-}
-
-private struct WorkflowDirectoryIdentity {
-    let manifest: WorkflowOutputDirectoryManifest
-    let sizeBytes: Int64
-    let sha256: String
 }
 
 private struct WorkflowPluginNodePreflight: Codable {
@@ -2367,33 +1285,8 @@ private struct WorkflowPluginNodeEvent: Codable {
     }
 }
 
-private func workflowValue(_ value: WorkflowValue, matches type: WorkflowPortType) -> Bool {
-    switch type {
-    case .string, .enumeration, .asset, .assetDirectory:
-        value.stringValue != nil
-    case .integer:
-        value.integerValue != nil
-    case .number:
-        value.numberValue != nil
-    case .boolean:
-        value.booleanValue != nil
-    case .json:
-        true
-    case .assetCollection, .assetArray:
-        if case .array = value { true } else { false }
-    }
-}
-
 func workflowNodeAllowsResumeReuse(_ node: WorkflowNode) -> Bool {
     node.arguments.values.allSatisfy { $0.secretNames.isEmpty }
 }
 
-private func isConfinedRelativeWorkflowPath(_ path: String) -> Bool {
-    !path.isEmpty
-        && !path.hasPrefix("/")
-        && path.split(separator: "/", omittingEmptySubsequences: false).allSatisfy { component in
-            !component.isEmpty && component != "." && component != ".."
-        }
-}
-
-private struct WorkflowCancellationError: Error {}
+struct WorkflowCancellationError: Error {}
